+++ packages/is/is.bigint/src/lib/isBigInt.ts
@@ -1,0 +1,3 @@
+export const isBigInt = (value: any): boolean => {
+  return typeof value === 'bigint';
+};