+++ packages/is/is.bigint/src/lib/isBigint.ts
@@ -1,3 +1,0 @@
-export const isBigint = (value: any): boolean => {
-  return typeof value === 'bigint';
-};