+++ packages/is/is.iterable/src/lib/isIterable.ts
@@ -1,0 +1,14 @@
+import { isNil } from '@zerodep/is.nil';
+
+export function isIterable(value: any): boolean {
+  // we do not want any kind of number or string to be iterable
+  const type = typeof value;
+  if (type === 'string' || isNil(value)) {
+    return false;
+  }
+
+  return (
+    typeof value[Symbol.iterator] === 'function' ||
+    typeof value[Symbol.asyncIterator] === 'function'
+  );
+}