+++ packages/is/is.weakset/rollup.config.ts
@@ -2,6 +2,11 @@ import ts from 'rollup-plugin-ts';
 
 const dir = 'is/is.weakset';
 
+const banner = `/**
+ * @source the https://github.com/cdepage/zerodep/tree/main/packages/${dir}
+ * @copyright Chris dePage
+ */`;
+
 export default [
   {
     input: `packages/${dir}/src/index.ts`,
@@ -9,12 +14,12 @@ export default [
       {
         file: `dist/packages/${dir}/index.esm.js`,
         format: 'esm',
-        banner: `/* @source the https://github.com/cdepage/zerodep/tree/main/packages/${dir} */`,
+        banner,
       },
       {
         file: `dist/packages/${dir}/index.cjs.js`,
         format: 'cjs',
-        banner: `/* @source the https://github.com/cdepage/zerodep/tree/main/packages/${dir} */`,
+        banner,
         interop: 'auto',
       },
     ],