+++ packages/to/to/src/index.spec.ts
@@ -1,0 +1,7 @@
+import * as converters from './index';
+
+describe('Guards barrel package', () => {
+  it('should export specific packages and interfaces', () => {
+    expect(Object.keys(converters)).toEqual(['toString']);
+  });
+});