+++ packages/is/is/CHANGELOG.md
@@ -4,20 +4,30 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Introduce the `is.iterable` utility
+- Improve README with better structure and examples
+- Add link to source code to the minified, distributed packages
+- Add more types and constructs to unit tests
+
 ## [0.3.x] - 2022-04-25
 
 ### Added
 
-Added isNil check
+- Introduce the `is.nil` utility
 
 ## [0.2.x] - 2022-04-25
 
 ### Added
 
-Added the following checks: isArray, isBoolean, isBigint, isDate
+- Introduce the `is.array` utility
+- Introduce the `is.boolean` utility
+- Introduce the `is.bigint` utility
+- Introduce the `is.date` utility
 
 ## [0.1.x] - 2022-04-25
 
 ### Added
 
-- "is" barrel package
+- Introduce the `is` barrel package