+++ packages/utils/CHANGELOG.md
@@ -4,24 +4,39 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Introduce the `is.iterable` utility
+
+### Changed
+
+- Improve README with better structure and examples
+- Add link to source code to the minified, distributed packages
+- Add more types and constructs to unit tests
+
 ## [0.3.x] - 2022-04-26
 
 ### Added
 
-Added isNil check
+- Introduce the `is.nil` utility
 
 ## [0.2.x] - 2022-04-26
 
 ### Added
 
-- Added guardBoolean
-- Added guardBigint
-- Added guardDate
-- Added the "is" suite of checks
+- Introduce the `guard.boolean` utility
+- Introduce the `guard.bigint` utility
+- Introduce the `guard.date` utility
+- Introduce the `is` barrel package
 
 ## [0.1.0] - 2022-04-25
 
 ### Added
 
-- ZeroDepError base error class
-- Guards for strings and numbers
+- Introduce the `guard` barrel package
+- Introduce the `guard.string` utility
+- Introduce the `guard.float` utility
+- Introduce the `guard.integer` utility
+- Introduce the `guard.number` utility